+++ gradle/libs.versions.toml
@@ -5,7 +5,7 @@ agp = "8.2.2"
 androidxVersion = "1.0.0"
 uberJava = "0.8.5"
 mavenPublish = "0.27.0"
-kotlin = "1.9.22"
+kotlin = "1.9.23"
 junit = "4.13.2"
 runner = "1.0.2"
 espresso-core = "3.0.2"
@@ -28,7 +28,11 @@ mockito = "5.11.0"
 mockito-kotlin = "5.2.1"
 kotlin-coroutines-test = "1.8.0"
 moshi = "1.15.0"
-androidxComposeBom = "2024.03.00"
+constraintlayout = "2.1.4"
+lifecycle-runtime-ktx = "2.7.0"
+activity-compose = "1.8.2"
+compose-bom = "2023.08.00"
+androidx-ui-tooling = "1.6.7"
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
@@ -65,6 +69,11 @@ androidx-test-ext-junit = { group = "androidx.test.ext", name = "junit", version
 androidx-test-espresso-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "androidx-test-espresso-espresso-core" }
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 kotlin-coroutines-test = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-test", version.ref = "kotlin-coroutines-test" }
-androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "androidxComposeBom" }
 androidx-compose-foundation = { group = "androidx.compose.foundation", name = "foundation" }
-androidx-compose-material3 = { group = "androidx.compose.material3", name = "material3" }
+lifecycle-runtime-ktx = { group = "androidx.lifecycle", name = "lifecycle-runtime-ktx", version.ref = "lifecycle-runtime-ktx" }
+activity-compose = { group = "androidx.activity", name = "activity-compose", version.ref = "activity-compose" }
+compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "compose-bom" }
+material3 = { group = "androidx.compose.material3", name = "material3" }
+androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling"}
+androidx-ui-tooling-preview-android = { group = "androidx.compose.ui", name = "ui-tooling-preview-android"}
+
