+++ settings.gradle.kts
@@ -45,6 +45,7 @@ include(
   ":core",
   ":core-android",
   ":rides-android",
+  ":samples:auth-demo",
   ":samples:request-button-sample",
   ":samples:login-sample",
   ":samples:login-with-auth-code-demo",
